requirements.txt
google-generativeai==0.8.3
streamlit==1.41.1
duckduckgo-search==6.3.7
python-dotenv==1.0.0
phidata
